All Skills

+++ Assets/Scripts/AIManager.cs
@@ -40,7 +40,7 @@ public class AIManager : MonoBehaviour
         {
             HexUnit userChess = DetectChess[i];
             //计算分数加权的方式，暂不考虑sp过三的事
-            int score = Mathf.Min(aiChess.UnitAttribute.Att - userChess.UnitAttribute.Def, userChess.UnitAttribute.hp) - Mathf.Min(userChess.UnitAttribute.Att - aiChess.UnitAttribute.Def, aiChess.UnitAttribute.hp);
+            int score = Mathf.Min(aiChess.UnitAttribute.Att - userChess.UnitAttribute.Def, (int)userChess.UnitAttribute.hp) - Mathf.Min(userChess.UnitAttribute.Att - aiChess.UnitAttribute.Def, (int)aiChess.UnitAttribute.hp);
             if (score > maxScore)
             {
                 maxScore = score;
